Rename source code files for DP SDK.
