+++ sdk/CMakeLists.txt
@@ -27,11 +27,8 @@ set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
 set(CMAKE_INSTALL_RPATH ${CMAKE_INSTALL_PREFIX}/lib)
 set(CMAKE_INSTALL_IPATH ${CMAKE_INSTALL_PREFIX}/include)
 
-add_library(mcm_dp SHARED
-    ${MCMDP_SRC_DIR}/mcm_dp.c
-    ${MCMDP_SRC_DIR}/impl_udp.c
-    ${MCMDP_SRC_DIR}/impl_memif.c
-    ${MCMDP_SRC_DIR}/media_proxy_ctrl.c)
+aux_source_directory(${MCMDP_SRC_DIR} sdk_srcs)
+add_library(mcm_dp SHARED ${sdk_srcs})
 
 target_include_directories(mcm_dp PUBLIC
     ${MCMDP_INCLUDE_DIR}
