+++ sdk/include/memif_impl.h
